Move redact tests to the redact package
